requirements.txt
# Package dependencies
pandas
numpy
pyyaml
jupyter
scikit-learn
torchvision
torch
matplotlib
scikit-image
loguru
streamlit